requirements.txt
# Core dependencies
numpy>=1.21.4
opencv-python>=4.5.5
opencv-contrib-python>=4.5.5
flask>=2.0.0
requests>=2.25.0
Pillow>=8.0.0

# Computer Vision and AI
depthai>=2.20.0
roboflow>=1.1.0
inference>=0.9.0

# Data processing
tqdm>=4.60.0
matplotlib>=3.3.0
scipy>=1.7.0

# ROS2 Python dependencies (if not using system packages)
# rclpy>=3.3.0
# sensor-msgs>=4.2.0
# geometry-msgs>=4.2.0
# std-msgs>=4.2.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipywidgets>=7.6.0

# System utilities
psutil>=5.8.0
pyyaml>=5.4.0
